requirements.txt
certifi==2018.4.16
decorator==4.0.11
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.2
django-static-sitemaps==4.2.1
django-tastypie==0.14.2
gunicorn==19.8.1
ipython==5.2.2
ipython-genutils==0.1.0
lxml==3.7.3
netifaces==0.10.5
olefile==0.45.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
pipenv==2018.7.1
prompt-toolkit==1.0.13
psycopg2==2.7.5
ptyprocess==0.5.1
Pygments==2.2.0
python-dateutil==2.6.0
python-mimeparse==1.6.0
pytz==2018.4
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.3.0
traitlets==4.3.1
virtualenv==16.0.0
virtualenv-clone==0.3.0
wcwidth==0.1.7
whitenoise==3.3.1
